source/Python_Superresolution/RoboyResolution/requirements.txt
matplotlib~=3.3.1
numpy~=1.19.1
boto3~=1.14.56
botocore~=1.17.56
tensorflow-gpu~=2.3.0
tensorflow-addons~=0.11.2
Pillow~=7.2.0
scikit-image~=0.17.2
scipy~=1.4.1
six~=1.15.0
opencv-python~=4.4.0.42
ipython~=7.17.0